partial splitting work
